+++ src/Exceptions/NotFoundException.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace CloudCreativity\LaravelJsonApi\Exceptions;
+
+use Exception;
+use Illuminate\Http\Response;
+use Neomerx\JsonApi\Exceptions\JsonApiException;
+
+class NotFoundException extends JsonApiException
+{
+
+    /**
+     * NotFoundException constructor.
+     *
+     * @param mixed $errors
+     * @param Exception|null $previous
+     */
+    public function __construct($errors = [], Exception $previous = null)
+    {
+        parent::__construct($errors, Response::HTTP_NOT_FOUND, $previous);
+    }
+}