+++ CHANGELOG.md
@@ -50,6 +50,8 @@ Last page link is now excluded if there are pages, rather than linking to page z
 Pagination meta will no longer leak into error response if error occurs when encoding data.
 - [#111](https://github.com/cloudcreativity/laravel-json-api/issues/111)
 Sending an invalid content type header now returns a JSON API error object.
+- [#146](https://github.com/cloudcreativity/laravel-json-api/issues/146)
+Return a 404 JSON API error object and allow this to be overridden.
 
 ## [0.12.0] - 2018-02-08